Resolve FIXME: SchemaVisitor/PropertyVisitor now get the SchemaTree, not just a JsonNode
This creates a new question: the SchemaTree can in theory tell anyone where it is, so
the whole explicit elementUri wouldn't be needed (as clearly visible in the code calculating it).
But: that code is ugly, and from the perspective of the user working with a pre-made URI is a lot
nicer in those visitors.

This might get resolved later with a utility method, or some API change in the json-schema classes.